+++ joulescope_ui/range_tool.py
@@ -17,9 +17,10 @@
 The tool is executed from a separate thread.
 """
 
-from PySide2 import QtWidgets, QtCore
+from PySide2 import QtCore
 import threading
 import time
+from queue import Queue, Empty
 import logging
 log = logging.getLogger(__name__)
 
@@ -27,29 +28,6 @@ log = logging.getLogger(__name__)
 SAMPLES_PER_ITERATION_DEFAULT_MIN = 1000
 PROGRESS_COUNT = 1000
 PROGRESS_UPDATE_RATE = 0.100  # seconds
-
-
-class Worker(QtCore.QObject):
-    sigFinished = QtCore.Signal(str)
-
-    def __init__(self, invoker):
-        super().__init__()
-        self._invoker = invoker
-        self._cancel = False
-
-    @QtCore.Slot()
-    def cancel(self):
-        self._invoker.stop()
-
-    def run(self):
-        try:
-            msg = self._invoker._thread_run()
-        except:
-            log.exception('During range tool run()')
-            msg = 'Error during range tool run()'
-        if msg is None:
-            msg = ''
-        self.sigFinished.emit(msg)
 
 
 class RangeToolIterable:
@@ -64,10 +42,9 @@ class RangeToolIterable:
         return self
 
     def __next__(self):
-        if self._x_next >= self._x_stop:
+        if self._x_next >= self._x_stop or self._parent.is_cancelled:
+            self._parent = None
             raise StopIteration()
-        if self._parent._cancel:
-            raise RuntimeError('Range tool canceled by user')
         self._parent.progress((self._x_next - self._x_start) / (self._x_stop - self._x_start))
         x_next = self._x_next + self._samples_per_iteration
         if x_next > self._x_stop:
@@ -81,18 +58,19 @@ class RangeToolInvoke(QtCore.QObject):  # also implements RangeToolInvocation
     sigProgress = QtCore.Signal(int)
     sigFinished = QtCore.Signal(object, str)  # range_tool, error message or ''
 
-    def __init__(self, parent, range_tool, cmdp):
+    def __init__(self, parent, resync_handler, range_tool, cmdp):
         super().__init__(parent)
         self._parent = parent
+        self._main_thread = threading.current_thread()
+        self._qt_resync_handler = resync_handler
         self._range_tool = range_tool
         self.cmdp = cmdp
         self._range_tool_obj = None
         self.sample_range = None
         self._time_range = None
         self._view = None
-        self._worker = None
         self._thread = None
-        self._progress = None
+        self._message_queue = Queue()
         self._cancel = False
         self._progress_time_last = time.time()
 
@@ -116,10 +94,21 @@ class RangeToolInvoke(QtCore.QObject):  # also implements RangeToolInvocation
             self._iterable = None
             raise
 
+    @property
+    def is_cancelled(self):
+        return self._cancel
+
     def samples_get(self):
         return self._view.samples_get(*self.sample_range, units='samples')
 
+    def _assert_worker_thread(self):
+        assert (threading.current_thread() == self._thread)
+
+    def _assert_main_thread(self):
+        assert (threading.current_thread() == self._main_thread)
+
     def iterate(self, samples_per_iteration=None):
+        self._assert_worker_thread()
         if samples_per_iteration is None or samples_per_iteration <= 0:
             if self.sample_frequency < SAMPLES_PER_ITERATION_DEFAULT_MIN:
                 samples_per_iteration = SAMPLES_PER_ITERATION_DEFAULT_MIN
@@ -129,10 +118,17 @@ class RangeToolInvoke(QtCore.QObject):  # also implements RangeToolInvocation
             samples_per_iteration = int(samples_per_iteration)
         return RangeToolIterable(self, samples_per_iteration)
 
+    def _qt_resync(self, cmd, args):
+        self._assert_worker_thread()
+        self._message_queue.put((cmd, args))
+        self._qt_resync_handler()  # causes self.on_resync()
+
     def progress(self, fraction):
+        self._assert_worker_thread()
         current_time = time.time()
         if current_time - self._progress_time_last > PROGRESS_UPDATE_RATE:
-            self.sigProgress.emit(int(fraction * PROGRESS_COUNT))
+            value = int(fraction * PROGRESS_COUNT)
+            self._qt_resync('progress', value)
             self._progress_time_last = current_time
 
     def _x_map_to_parent(self, x):
@@ -155,11 +151,12 @@ class RangeToolInvoke(QtCore.QObject):  # also implements RangeToolInvocation
     def run(self, view, statistics, x_start, x_stop):
         """Export data request.
 
-        :param view: The view implementation with TBD API.
+        :param view: The view implementation.
         :param statistics: The statistics (see :meth:`joulescope.driver.statistics_get`).
         :param x_start: The starting position in x-axis units.
         :param x_stop: The stopping position in x-axis units.
         """
+        self._assert_main_thread()
         self.statistics = statistics
         t1, t2 = min(x_start, x_stop), max(x_start, x_stop)
         log.info('range_tool %s(%s, %s)', self._range_tool.name, t1, t2)
@@ -179,53 +176,73 @@ class RangeToolInvoke(QtCore.QObject):  # also implements RangeToolInvocation
             if hasattr(self._range_tool_obj, 'run_pre'):
                 rc = self._range_tool_obj.run_pre(self)
                 if rc is not None:
-                    log.info('%s run_pre failed: %s', self._range_tool.name, rc)
+                    log.warning('%s run_pre failed: %s', self._range_tool.name, rc)
+                    self._finalize(f'{self._range_tool.name}: {rc}')
                     return
         except:
             log.exception('During range tool run_pre()')
             return
-
-        title = f'{self._range_tool.name} in progress...'
-        self._progress = QtWidgets.QProgressDialog(title, 'Cancel', 0, PROGRESS_COUNT, self._parent)
-        self._progress.setWindowTitle('Progress')
-        self._progress.setWindowModality(QtCore.Qt.WindowModal)
-        self._progress.setMinimumDuration(0)
-        self._worker = Worker(self)
-        self._worker.sigFinished.connect(self._on_finished, type=QtCore.Qt.QueuedConnection)
-        self._worker.moveToThread(self._thread)
-        self._thread = threading.Thread(target=self._worker.run)
-        self._progress.canceled.connect(self._on_cancel, type=QtCore.Qt.QueuedConnection)
-        self.sigProgress.connect(self._progress.setValue, type=QtCore.Qt.QueuedConnection)
+        self._thread = threading.Thread(target=self._thread_run)
         self._thread.start()
 
     def _thread_run(self):
-        return self._range_tool_obj.run(self)
+        self._assert_worker_thread()
+        try:
+            rv = self._range_tool_obj.run(self)
+            if self.is_cancelled:
+                rv = f'{self._range_tool.name}: Cancelled'
+        except Exception as ex:
+            log.exception('range tool run exception')
+            rv = f'{self._range_tool.name}: ERROR'
+        self._qt_resync('done', rv)
 
-    def _on_cancel(self):
+    def on_resync(self):
+        self._assert_main_thread()  # indirectly by self._qt_resync_callback
+        while True:
+            try:
+                cmd, args = self._message_queue.get(timeout=0.0)
+            except Empty:
+                break
+            except:
+                log.exception('on_resync message_queue get')
+            if cmd == 'progress':
+                self.sigProgress.emit(args)
+            elif cmd == 'done':
+                self._on_finished(args)
+
+    @QtCore.Slot()
+    def on_cancel(self):
+        log.info('range tool cancelled by user')
         self._cancel = True
 
     def _on_finished(self, msg):
-        self._thread.join()
+        self._assert_main_thread()
+        self.sigProgress.emit(1000)
         self.sigProgress.disconnect()
-        self._progress.hide()
-        self._progress.close()
-        time.sleep(0.0)  # yield to let progress window close
+        self._thread.join()
         self._thread = None
-        self._progress = None
-        self._worker = None
-
-        try:
-            if hasattr(self._range_tool_obj, 'run_post'):
-                self._range_tool_obj.run_post(self)
-        except:
-            log.exception('During range tool run_post()')
-            return
-
-        while len(self._commands):
-            command = self._commands.pop(0)
+        if not self.is_cancelled:
             try:
-                command()
+                if hasattr(self._range_tool_obj, 'run_post'):
+                    self._range_tool_obj.run_post(self)
             except:
-                log.exception('During range tool command')
+                log.exception('During range tool run_post()')
+                return
 
+            while not self.is_cancelled and len(self._commands):
+                command = self._commands.pop(0)
+                try:
+                    command()
+                except:
+                    log.exception('During range tool command')
+        self._finalize(msg)
+
+    def _finalize(self, msg):
+        self._assert_main_thread()
+        log.info('range tool finalize')
         self.sigFinished.emit(self._range_tool, msg)
+        self._range_tool_obj = None
+        self._parent = None
+        self._qt_resync_handler = None
+        self._range_tool = None
+        self.cmdp = None